resolved PR changes
